% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{datatype}
\alias{datatype}
\title{datatype}
\usage{
datatype(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A data type string
}
\description{
"Bags" R vector data types (from class) into three bins:
num = {integer, numeric}
char = {character, factor}
bool = {logical}
Returns binned data type
}
\examples{
datatype(c(1,2,3)) == datatype(c(1L,2L,3L))
[1] TRUE
datatype(c('a','b')) == datatype(factor('a','b'))
[1] TRUE
}
