% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAlterFile}
\alias{getAlterFile}
\title{getAlterFile}
\usage{
getAlterFile(dir, clade)
}
\arguments{
\item{dir}{The directory to look for the file}
}
\value{
a list with names header and fun
}
\description{
Reads a clade object from a procedure file in a directory
The file is chosen from filename.
}
\examples{
mtcarsAlter <- getCladeFile("clades","mtcars")
}
