% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAlterObject}
\alias{getAlterObject}
\title{getAlterObject}
\usage{
getAlterObject(file)
}
\arguments{
\item{file}{The file to read}
}
\value{
A list with names header and fun
}
\description{
Get an alter object from a file, separating and reading the yaml header,
and sourcing the r code with dget.
}
\examples{
getAlter("mtcars.R") 
}
