% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readDispatch}
\alias{readDispatch}
\title{readDispatch}
\usage{
readDispatch(file)
}
\arguments{
\item{file}{A file with an extension}
}
\value{
a data-table or R object
}
\description{
Read a file according to its extension
}
\examples{
data <- readDispatch('somedata.csv')
}
