% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readSpec}
\alias{readSpec}
\title{readSpec}
\usage{
readSpec(file)
}
\arguments{
\item{file}{The file to read}
}
\value{
A list with names header and fun
}
\description{
Get a spec from a file, separating and reading the yaml header, and sourcing
the r code with dget.
}
\examples{
readSpec("mtcars.R") 
}
