% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codder.R
\name{safeTransform}
\alias{safeTransform}
\title{safeTransform}
\usage{
safeTransform(data, alterlist)
}
\arguments{
\item{data}{data-table or tibble}

\item{alterlist}{a list with names header and fun}
}
\value{
A transformed tibble
}
\description{
Transform a table using a cladefile, specifying how the
data is to be transformed, and what assumptions should
be fulfilled.
}
\examples{
mtcars_fixed <- safeTransform(mtcars, 'clades/mtcars.R')
}
