% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testColumns}
\alias{testColumns}
\title{testColumns}
\usage{
testColumns(data, schema)
}
\arguments{
\item{data}{A data frame}

\item{data}{A schema}
}
\value{
TRUE if passed, FALSE if not
}
\description{
Perform tests on the columns of a data frame, to see if
it corresponds with a given schema
}
