% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testEqual}
\alias{testEqual}
\title{testEqual}
\usage{
testEqual(x, y, msg = "something")
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{msg}{What is being tested}
}
\value{
TRUE if passed, FALSE if not
}
\description{
A verbose function that checks if x and y are equal
}
\examples{
testEqual(c(1,2,3),c(1,2,3),msg = 'simple test')
}
