% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testSchema}
\alias{testSchema}
\title{testSchema}
\usage{
testSchema(data, schema)
}
\arguments{
\item{data}{A data frame}

\item{data}{A schema}
}
\value{
TRUE if passed, FALSE if not
}
\description{
Apply lots of tests to a data frame
}
