% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyTransform.R
\name{tidyTransform}
\alias{tidyTransform}
\title{tidyTransform}
\usage{
tidyTransform(data, spec)
}
\arguments{
\item{data}{data-table or tibble}

\item{spec}{a list with names header and fun}
}
\value{
A transformed tibble
}
\description{
Transform a table using a spec, specifying how the
data is to be transformed, and what assumptions should
be fulfilled.
}
\examples{
mtcars_fixed <- safeTransform(mtcars, 'specs/mtcars.R')
}
