% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validateColumns}
\alias{validateColumns}
\title{validateColumns 
 
Asserts that: 
Columns have the column names specified in the schema
Columns have the data types specified in the schema}
\usage{
validateColumns(data, spec)
}
\arguments{
\item{data}{A data frame}

\item{spec}{A spec}
}
\description{
validateColumns 
 
Asserts that: 
Columns have the column names specified in the schema
Columns have the data types specified in the schema
}
