% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validateData}
\alias{validateData}
\title{validateData}
\usage{
validateData(data, spec, type = "input")
}
\arguments{
\item{data}{A data frame}

\item{spec}{A spec}

\item{type}{A Either 'input' or 'output', to reserve some validation for either}
}
\value{
TRUE if passed, FALSE if not
}
\description{
Apply lots of tests to a data frame
}
